test_driver: introduce testDomainJobWait
This patch adds function testDomainJobWait. Since the test_driver
exists in the application's address space, it is impossible to test the
scenario where an application starts a core dump job, restarts, and then
joins it. To get around this problem, we simulate a dump job by having
testDomainGetJobStats check for the LIBVIRT_TEST_DRIVER_DUMP_JOB
environment variable and it's value used by testDomainJobWait as
the return value of the dump job.

Signed-off-by: Thanos Makatos <thanos.makatos@nutanix.com>